requirement.txt
alembic==1.0.10
asn1crypto==0.24.0
Babel==2.6.0
bleach==3.1.0
blinker==1.4
cffi==1.12.3
Click==7.0
cryptography==2.6.1
cymysql==0.9.13
Flask==1.0.3
Flask-Admin==1.5.3
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.1
Flask-Script==2.0.6
Flask-SocketIO==4.1.0
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==6.0.0
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.8.1
python-socketio==4.1.0
pytz==2019.1
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.3
webencodings==0.5.1
Werkzeug==0.15.4
WTForms==2.2.1
